src/ocr-serving/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pillow
opencv-python-headless